pyproject.toml
[project]
name = "video-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=12.1.0",
    "yt-dlp>=2025.12.8",
]
